requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
click==8.1.7
    # via uvicorn
dnspython==2.4.2
    # via
    #   email-validator
    #   pymongo
email-validator==2.1.0.post1
    # via pydantic
fastapi==0.104.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   -r requirements.in
    #   uvicorn
idna==3.4
    # via
    #   anyio
    #   email-validator
motor==3.3.2
    # via -r requirements.in
phonenumbers==8.13.25
    # via -r requirements.in
pydantic[email]==2.4.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.10.1
    # via pydantic
pymongo==4.6.0
    # via motor
python-dotenv==1.0.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.24.0.post1
    # via -r requirements.in
